+++ config/config.default.ts
@@ -74,7 +74,11 @@ export default (appInfo: EggAppConfig) => {
     user: process.env.CNPMCORE_MYSQL_USER || process.env.MYSQL_USER || 'root',
     password: process.env.CNPMCORE_MYSQL_PASSWORD || process.env.MYSQL_PASSWORD,
     charset: 'utf8mb4',
-    logger: {},
+    logger: {
+      // https://github.com/cyjake/leoric/blob/master/docs/zh/logging.md#logqueryerror
+      // ignore query error
+      logQueryError() {},
+    },
   };
 
   config.redis = {
@@ -98,6 +102,8 @@ export default (appInfo: EggAppConfig) => {
       return ctx.get('Origin');
     },
     credentials: true,
+    // https://github.com/koajs/cors/blob/master/index.js#L10C57-L10C64
+    allowMethods: 'GET,HEAD,PUT,POST,DELETE,PATCH,OPTIONS',
   };
 
   config.nfs = {